deployment/requirements_gpu.txt
# GPU-specific requirements for RunPod deployment
# This file ensures all dependencies work with CUDA

# PyTorch with CUDA 11.7 support
--index-url https://download.pytorch.org/whl/cu117
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# Model architectures (CUDA-enabled)
pykan>=0.2.0
mamba-ssm>=1.2.0
git+https://github.com/Blealtan/efficient-kan.git

# Training framework
pytorch-lightning>=2.0.0
wandb>=0.15.0

# ML and data processing
transformers>=4.30.0
datasets>=2.14.0
evaluate>=0.4.0
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0
scikit-learn>=1.3.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Jupyter and interactive tools
jupyter>=1.0.0
jupyterlab>=4.0.0
ipywidgets>=8.0.0
notebook>=6.0.0

# Monitoring and utilities
psutil>=5.9.0
python-dotenv>=1.0.0
requests>=2.31.0
tqdm>=4.65.0
rich>=13.0.0

# Development tools
pytest>=7.0.0
black>=23.0.0
isort>=5.12.0

# GPU monitoring
pynvml>=11.5.0

# Optional: Advanced profiling (install manually if needed)
# fvcore  # For FLOP counting
# memory-profiler  # For memory profiling
# tensorboard  # For logging and visualization